pyproject.toml
# Copyright (c) Meta Platforms, Inc. and affiliates.
# All rights reserved.
#
# This source code is licensed under the BSD-style license found in the
# LICENSE file in the root directory of this source tree.

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-os_expert_env"
version = "0.1.0"
description = "OS Expert Environment — high-fidelity Linux sandbox for RL agent training"
requires-python = ">=3.10"
dependencies = [
    # Core OpenEnv runtime (provides FastAPI server + HTTP client types)
    "openenv-core[core]>=0.2.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
# Server entry point - enables running via: uv run --project . server
# or: python -m os_expert_env.server.app
server = "os_expert_env.server.app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["os_expert_env*", "server*", "env*"]

[tool.setuptools.package-dir]
"os_expert_env" = "."
"os_expert_env.server" = "server"
"os_expert_env.env" = "env"
"os_expert_env.env.tools" = "env/tools"